requirements.txt
# Core dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-xdist>=3.5.0
pytest-html>=4.1.1
playwright>=1.40.0
aiohttp>=3.9.0
pandas>=2.1.0
numpy>=1.26.0
jinja2>=3.1.2
PyYAML>=6.0.1

# Optional dependencies
allure-pytest>=2.13.0
requests>=2.31.0
influxdb-client>=1.36.0

# Agent system dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
paramiko>=3.3.1
asyncssh>=2.14.0
click>=8.1.7

# Development dependencies
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
